+++ src/types/eslint-plugin-import.d.ts
@@ -1,9 +1,0 @@
-declare module "eslint-plugin-import" {
-  type ImportPlugin = {
-    flatConfigs: {
-      recommended: import("typescript-eslint").InfiniteDepthConfigWithExtends;
-    };
-  };
-  const importPlugin: ImportPlugin;
-  export = importPlugin;
-}